Split address logs tests
