+++ tests/tools/test_address_logs.py
@@ -1,0 +1,250 @@
+# tests/tools/test_address_logs.py
+import pytest
+from unittest.mock import patch, AsyncMock, MagicMock
+import httpx
+import json
+
+from blockscout_mcp_server.tools.address_tools import get_address_logs
+from blockscout_mcp_server.tools.common import encode_cursor
+
+@pytest.mark.asyncio
+async def test_get_address_logs_success(mock_ctx):
+    """
+    Verify get_address_logs correctly processes and formats address logs.
+    """
+    # ARRANGE
+    chain_id = "1"
+    address = "0x123abc"
+    mock_base_url = "https://eth.blockscout.com"
+
+    mock_api_response = {
+        "items": [
+            {
+                "address": address,
+                "topics": ["0xtopic1...", "0xtopic2..."],
+                "data": "0xdata123...",
+                "log_index": "0",
+                "transaction_hash": "0xtx123...",
+                "block_number": 19000000
+            }
+        ]
+    }
+
+    with patch('blockscout_mcp_server.tools.address_tools.get_blockscout_base_url', new_callable=AsyncMock) as mock_get_url, \
+         patch('blockscout_mcp_server.tools.address_tools.make_blockscout_request', new_callable=AsyncMock) as mock_request:
+
+        mock_get_url.return_value = mock_base_url
+        mock_request.return_value = mock_api_response
+
+        # ACT
+        result = await get_address_logs(chain_id=chain_id, address=address, ctx=mock_ctx)
+
+        # ASSERT
+        mock_get_url.assert_called_once_with(chain_id)
+        mock_request.assert_called_once_with(
+            base_url=mock_base_url,
+            api_path=f"/api/v2/addresses/{address}/logs",
+            params={}
+        )
+        
+        # Verify the result starts with the expected prefix
+        expected_prefix = "**Items Structure:**"
+        assert result.startswith(expected_prefix)
+        
+        # Verify the JSON content is included
+        assert f'"address": "{address}"' in result
+        assert '"0xtopic1..."' in result
+        assert '"0xtopic2..."' in result
+        
+        assert mock_ctx.report_progress.call_count == 3
+        assert mock_ctx.info.call_count == 3
+
+@pytest.mark.asyncio
+async def test_get_address_logs_with_pagination(mock_ctx):
+    """Verify get_address_logs includes pagination hint and correctly formats the main JSON body."""
+    chain_id = "1"
+    address = "0x123abc"
+    mock_base_url = "https://eth.blockscout.com"
+
+    mock_api_response = {
+        "items": [
+            {
+                "address": address,
+                "topics": ["0xtopic1..."],
+                "data": "0xdata123...",
+                "log_index": "0",
+                "transaction_hash": "0xtx123...",
+                "block_number": 19000000
+            }
+        ],
+        "next_page_params": {
+            "block_number": 18999999,
+            "index": "42",
+            "items_count": 50
+        }
+    }
+
+    fake_cursor = "ENCODED_CURSOR_STRING_FROM_TEST"
+    fake_json_body = '{"items": [...], "next_page_params": {...}}'
+
+    with patch('blockscout_mcp_server.tools.address_tools.get_blockscout_base_url', new_callable=AsyncMock) as mock_get_url, \
+         patch('blockscout_mcp_server.tools.address_tools.make_blockscout_request', new_callable=AsyncMock) as mock_request, \
+         patch('blockscout_mcp_server.tools.address_tools.json.dumps') as mock_json_dumps, \
+         patch('blockscout_mcp_server.tools.address_tools.encode_cursor') as mock_encode_cursor:
+
+        mock_get_url.return_value = mock_base_url
+        mock_request.return_value = mock_api_response
+        mock_json_dumps.return_value = fake_json_body
+        mock_encode_cursor.return_value = fake_cursor
+
+        result = await get_address_logs(chain_id=chain_id, address=address, ctx=mock_ctx)
+
+        mock_json_dumps.assert_called_once_with(mock_api_response)
+        mock_encode_cursor.assert_called_once_with(mock_api_response["next_page_params"])
+
+        assert result.startswith("**Items Structure:**")
+        assert fake_json_body in result
+        assert f'cursor="{fake_cursor}"' in result
+
+        mock_get_url.assert_called_once_with(chain_id)
+        mock_request.assert_called_once_with(
+            base_url=mock_base_url,
+            api_path=f"/api/v2/addresses/{address}/logs",
+            params={}
+        )
+        assert mock_ctx.report_progress.call_count == 3
+        assert mock_ctx.info.call_count == 3
+
+@pytest.mark.asyncio
+async def test_get_address_logs_with_optional_params(mock_ctx):
+    """
+    Verify get_address_logs correctly passes optional pagination parameters.
+    """
+    # ARRANGE
+    chain_id = "1"
+    address = "0x123abc"
+    block_number = 18999999
+    index = 42  # Fixed parameter name from log_index to index
+    items_count = 25
+    mock_base_url = "https://eth.blockscout.com"
+
+    mock_api_response = {"items": []}
+
+    with patch('blockscout_mcp_server.tools.address_tools.get_blockscout_base_url', new_callable=AsyncMock) as mock_get_url, \
+         patch('blockscout_mcp_server.tools.address_tools.make_blockscout_request', new_callable=AsyncMock) as mock_request:
+
+        mock_get_url.return_value = mock_base_url
+        mock_request.return_value = mock_api_response
+
+        # ACT
+        cursor = encode_cursor({"block_number": block_number, "index": index, "items_count": items_count})
+        result = await get_address_logs(
+            chain_id=chain_id,
+            address=address,
+            cursor=cursor,
+            ctx=mock_ctx
+        )
+
+        # ASSERT
+        mock_get_url.assert_called_once_with(chain_id)
+        mock_request.assert_called_once_with(
+            base_url=mock_base_url,
+            api_path=f"/api/v2/addresses/{address}/logs",
+            params={
+                "block_number": block_number,
+                "index": index,
+                "items_count": items_count,
+            }
+        )
+        
+        # Verify result structure
+        assert result.startswith("**Items Structure:**")
+        assert '"items": []' in result
+        
+        assert mock_ctx.report_progress.call_count == 3
+        assert mock_ctx.info.call_count == 3
+
+@pytest.mark.asyncio
+async def test_get_address_logs_invalid_cursor(mock_ctx):
+    """Verify the tool returns a user-friendly error for a bad cursor."""
+    chain_id = "1"
+    address = "0x123abc"
+    invalid_cursor = "bad-cursor"
+
+    result = await get_address_logs(chain_id=chain_id, address=address, cursor=invalid_cursor, ctx=mock_ctx)
+
+    assert "Error: Invalid or expired pagination cursor." in result
+
+@pytest.mark.asyncio
+async def test_get_address_logs_api_error(mock_ctx):
+    """
+    Verify get_address_logs correctly propagates API errors.
+    """
+    # ARRANGE
+    chain_id = "1"
+    address = "0x123abc"
+    mock_base_url = "https://eth.blockscout.com"
+
+    api_error = httpx.HTTPStatusError("Internal Server Error", request=MagicMock(), response=MagicMock(status_code=500))
+
+    with patch('blockscout_mcp_server.tools.address_tools.get_blockscout_base_url', new_callable=AsyncMock) as mock_get_url, \
+         patch('blockscout_mcp_server.tools.address_tools.make_blockscout_request', new_callable=AsyncMock) as mock_request:
+
+        mock_get_url.return_value = mock_base_url
+        mock_request.side_effect = api_error
+
+        # ACT & ASSERT
+        with pytest.raises(httpx.HTTPStatusError):
+            await get_address_logs(chain_id=chain_id, address=address, ctx=mock_ctx)
+
+        mock_get_url.assert_called_once_with(chain_id)
+        mock_request.assert_called_once_with(
+            base_url=mock_base_url,
+            api_path=f"/api/v2/addresses/{address}/logs",
+            params={}
+        )
+
+@pytest.mark.asyncio
+async def test_get_address_logs_empty_logs(mock_ctx):
+    """
+    Verify get_address_logs handles addresses with no logs.
+    """
+    # ARRANGE
+    chain_id = "1"
+    address = "0x123abc"
+    mock_base_url = "https://eth.blockscout.com"
+
+    mock_api_response = {"items": []}
+
+    # Patch json.dumps directly since it's imported locally in the function
+    with patch('blockscout_mcp_server.tools.address_tools.get_blockscout_base_url', new_callable=AsyncMock) as mock_get_url, \
+         patch('blockscout_mcp_server.tools.address_tools.make_blockscout_request', new_callable=AsyncMock) as mock_request, \
+         patch('blockscout_mcp_server.tools.address_tools.json.dumps') as mock_json_dumps:
+
+        mock_get_url.return_value = mock_base_url
+        mock_request.return_value = mock_api_response
+        # We don't care what json.dumps returns, only that it's called correctly
+        mock_json_dumps.return_value = "{...}"
+
+        # ACT
+        result = await get_address_logs(chain_id=chain_id, address=address, ctx=mock_ctx)
+
+        # ASSERT
+        # Assert that json.dumps was called with the exact API response data
+        mock_json_dumps.assert_called_once_with(mock_api_response)
+
+        mock_get_url.assert_called_once_with(chain_id)
+        mock_request.assert_called_once_with(
+            base_url=mock_base_url,
+            api_path=f"/api/v2/addresses/{address}/logs",
+            params={}
+        )
+        
+        # Verify the result structure
+        assert result.startswith("**Items Structure:**")
+        
+        # Verify no pagination hint is included
+        assert "To get the next page call" not in result
+        
+        assert mock_ctx.report_progress.call_count == 3
+        assert mock_ctx.info.call_count == 3
